Fixed loading weights from checkpoint
At head, the weights are loaded from the coco checkpoint, which changes the dimensions of some layers, then the model is loaded again from the intended checkpoint,
which changes back the dimensions of the layers. The result is that any layer that doesn't match the dimensions of the coco checkpoint cannot be loaded from
a previous checkpoint and is re-initialized instead.

With this change, if a checkpoint is provided, the model is loaded directly from it.

Test Plan:
- python simple_training.py --load my_checkpoint.pt
- Before the change, not all layers were loaded ("Transferred 355/397 items from pretrained weights")
- After the change, all layers are loaded.